Replace event context with request scope
