+++ src/lib.rs
@@ -1,10 +1,10 @@
 /// Pre-condition checks for use with [`async-graphql`](https://docs.rs/async-graphql)
 #[cfg(feature = "graphql")]
 pub mod checks;
-/// Context information from the `events` service
-pub mod event;
 #[cfg(feature = "extract")]
 mod headers;
+/// Context information from the `events` service
+pub mod scope;
 /// Context information from the `identity` service
 pub mod user;
